+++ Snowflake.Data/Core/Authenticator/OAuthCacheKeys.cs
@@ -20,14 +20,14 @@ namespace Snowflake.Data.Core.Authenticator
             if (IsCacheAvailableForAuthorizationCodeFlow(user, clientStoreTemporaryCredentials, true))
             {
                 accessTokenKey = SnowflakeCredentialManagerFactory.BuildCacheKey(new CacheKeyInput(
-                    tokenType: TokenType.OAuthAccessToken.GetAttribute<StringAttr>().value,
+                    tokenType: TokenType.OAuthAccessToken.ToCacheKeyPrefix(),
                     idp: idpUrl,
                     snowflake: snowflakeHost,
                     username: user,
                     role: role ?? string.Empty
                 ));
                 refreshTokenKey = SnowflakeCredentialManagerFactory.BuildCacheKey(new CacheKeyInput(
-                    tokenType: TokenType.OAuthRefreshToken.GetAttribute<StringAttr>().value,
+                    tokenType: TokenType.OAuthRefreshToken.ToCacheKeyPrefix(),
                     idp: idpUrl,
                     snowflake: snowflakeHost,
                     username: user,